Validaciones web

+++ web/src/components/works/WorkForm.jsx
@@ -2,7 +2,7 @@ import { Box, Button, Chip, FormControl, Grid, InputLabel, MenuItem, Select, Sta
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from 'react-redux';
 import Swal from "sweetalert2";
-import { createOneWork} from "../../redux/slices/workSlice"
+import { createOneWork, getAllWorks} from "../../redux/slices/workSlice"
 import NavDashboard2 from "../navDachboard2/NavDashboard2";
 import NavDashboard from "../navDashboard/NavDashboard";
 import { useNavigate } from "react-router-dom";
@@ -14,8 +14,20 @@ export default function WorkForm () {
   const dispatch = useDispatch();
   const categories = useSelector(state => state.categories.categories);
   const user = useSelector(state => state.users.list);
+  const work = useSelector(state => state.works.workList);
 
   const navigate = useNavigate();
+  
+  const [ arrayObras , setArrayObras ] = useState([])
+  const [ number, setNumbre ] = useState(0)
+  
+  if(work && number < 1 ){
+    work.map( (e) =>{
+      setArrayObras(state => [...state,e.name]);
+    })
+    setNumbre( number + 1 )
+  }
+
 
   const [ createWorkState, setCreateWorkState ] = useState({
     name: "" ,
@@ -43,6 +55,18 @@ export default function WorkForm () {
   const [ categoriaUnica , setCategoriaUnica] = useState({
     name:"name"
   })
+
+  const [ stateHeigthValues, setStateHeigthValues ] = useState(0)
+
+  // const [ number, setNumbre ] = useState(0)
+
+  // if(work && number < 1 ){
+  //     work.progresses.map((e) =>{
+  //       setStateHeigthValues(stateHeigthValues + e.height_value);
+  //     })
+  //     setNumbre( number + 1 )
+  //     console.log(stateHeigthValues)
+  // }
   
   const handleSelectCategoria = (e) =>{
     SetSelectedCategory(e.target.value);
@@ -55,7 +79,21 @@ export default function WorkForm () {
   }
 
 const deleteCategoria =  (name) =>{
-  dispatch(deleteOneCategory(name))
+  Swal.fire({
+    title: `¿Estás seguro que quieres borrar la categoria ${name}?`,
+    icon: 'question',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'si!'
+  }).then((borrado) => {
+    if (borrado.isConfirmed) {
+      dispatch(deleteOneCategory(name))
+      Swal.fire(
+        'categoria borrada!'
+      )
+    }
+  })
 } 
 
   const handleCategoriaChip = (e) =>{
@@ -74,11 +112,25 @@ const [ array , setArray] = useState([])
 
 const handleAdd = () =>{
 
+  if(selectedCategory.name === ""){
+    return Swal.fire({title: 'llene los campos para añadir categoria'})
+  }
+
+  if(progress.value === "" ){
+    return Swal.fire({title: 'llene el avance de la categoria'})
+  }else if(progress.height_value === ""){
+    return Swal.fire({title: 'llene el peso de la categoria'})
+  }
+
     if(!array.includes(selectedCategory)){
       setShip(state => [...state, {category: selectedCategory, progress: progress }]);
     }else{
-      console.log('no se puede agregar');
+      Swal.fire({title: 'categoria ya está añadida'})
     }
+    // setStateHeigthValues(stateHeigthValues + Number(progress.height_value));
+    // if(stateHeigthValues + progress.value > 100 ){
+    //     return  Swal.fire({title: 'no se puede crear'})
+    // }
     setArray([...array, selectedCategory])
   }
 
@@ -90,15 +142,19 @@ const handleAdd = () =>{
   const handleCreateCategoria  = (e) =>{
     let array = []
     e.preventDefault();
+    if(categoriachip == null || categoriachip === ''){
+      return Swal.fire({title: 'llene el campo para crear categoria'})
+    }
     categories.map( (e) =>{
       array.push( e.name)
+      Swal.fire({title: 'categoria creada'})
     })
     if(!array.includes(categoriachip)){
-      dispatch(createOneCategory(categoriachip))
+      dispatch(createOneCategory(categoriachip.trim().toLowerCase()))
       createOneCategory(categoriachip)
       dispatch(getAllCategories())
     }else{
-      console.log('no se creo');
+      Swal.fire({title: 'categoria ya existente'})
       dispatch(getAllCategories())
     }
     dispatch(getAllCategories())
@@ -107,6 +163,7 @@ const handleAdd = () =>{
   useEffect(() => {
     dispatch(getAllCategories())
     dispatch(getAllUsers())
+    dispatch(getAllWorks())
     setCreateWorkState('')
   }, [dispatch]);
 
@@ -116,12 +173,27 @@ const handleAdd = () =>{
     setCreateWorkState(state => ({...state, [name]: value}));
   }
 
+
+
+
   const createWork = async () => {
-    await dispatch(createOneWork({work: createWorkState, ships: ship, shipUsers: shipUsers}));
-    Swal.fire({
-      title: 'Obra creada!',
-    })
-    navigate('/work');
+
+    if(arrayObras.includes(createWorkState.name)){
+      return Swal.fire({
+        title: `la obra ${createWorkState.name} ya existe`,
+      })
+    }
+    
+    if(createWorkState ){
+      await dispatch(createOneWork({work: createWorkState, ships: ship, shipUsers: shipUsers}));
+      Swal.fire({
+        title: 'Obra creada!',
+      })
+      navigate('/work');
+    }else{
+      return Swal.fire({title: 'Llene los campos para crear una obra'})
+    }
+    
   }
 
 
@@ -227,6 +299,8 @@ return (
                       </Select>
                       </FormControl>
                       <TextField  
+                        type='number'
+                        InputProps={{inputProps:{min:10, max:100, step:5}}}
                         onChange={handleChange}
                         value={progress.value}
                         label="% Avance"
@@ -241,6 +315,7 @@ return (
                         boxShadow: '5px 5px 13px 2px rgba(0,0,0,0.39)'}} 
                         />
                       <TextField  
+                      type='number'
                       onChange={handleChange}
                       value={progress.height_value}
                       label="% Peso de la categoria"
@@ -251,9 +326,11 @@ return (
                                     fontSize:".8rem",
                                   }
                                 }} 
+                      InputProps={{inputProps:{min:10, max:100, step:5}}}
                       sx={{marginTop:'2rem',border:'none', 
                       boxShadow: '5px 5px 13px 2px rgba(0,0,0,0.39)'}} 
                       />
+                      
                       <Box sx={{marginTop:'2rem'}}>
                         <Button sx={{fontSize:'.7rem' , backgroundColor:'rgb(160, 7, 7) ', color:'#fff'}} onClick={() => handleAdd(selectedCategory.name)}>crear</Button>
                       </Box>
